src/transform-audio/requirements.txt
openai-whisper==20240930
torch==2.5.1